перевірка
