+++ src/partials/favorites.html
@@ -25,16 +25,16 @@
       </div>
 
       <picture>
-        <source srcset="./img/favorites/women_sport_desktop.jpg 1x, " media="(min-width: 1440px)">
-        <source srcset="./img/favorites/women_sport_desktop.webp 1x," media="(min-width: 1440px)" type="image/webp">
+        <source srcset="../img/favorites/women_sport_desktop.jpg 1x, " media="(min-width: 1440px)">
+        <source srcset="../img/favorites/women_sport_desktop.webp 1x," media="(min-width: 1440px)" type="image/webp">
 
-        <source srcset="./img/favorites/women_sport_tablet.jpg 1x, " media="(min-width: 768px)">
-        <source srcset="./img/favorites/women_sport_tablet.webp 1x," media="(min-width: 768px)" type="image/webp">
+        <source srcset="../img/favorites/women_sport_tablet.jpg 1x, " media="(min-width: 768px)">
+        <source srcset="../img/favorites/women_sport_tablet.webp 1x," media="(min-width: 768px)" type="image/webp">
 
-        <source srcset="./img/favorites/women_sport_mobile.jpg 1x, " media="(max-width: 767px)">
-        <source srcset="/./img/favorites/women_sport_mobile.webp" media="(max-width: 767px)" type="image/webp">
+        <source srcset="../img/favorites/women_sport_mobile.jpg 1x, " media="(max-width: 767px)">
+        <source srcset="../img/favorites/women_sport_mobile.webp" media="(max-width: 767px)" type="image/webp">
 
-        <img class="sports_women" src="/img/women_sports.jpg" alt="">
+        <img class="sports_women" src='../img/favorites/women_sports.jpg' alt="">
       </picture>
     </div>
